lists are now dragable, changed drag over and drag end to use draghelper, drag helper decides which drag method to use if dragging a list or a card. added ondragoverlist, ondragendlist, and applydraglist. removed log statements. position now updates correctly when a list is dragged similar how a card position is updated. updated list position field in django to be the same as card, anytime list position is updated board counter is also updated just like card. added logic to reset list positions to whole numbers just like card when board counter is reset. changed list serializer to be the same as card formats list position when fetched.

+++ kanban/models.py
@@ -12,7 +12,7 @@ class Board(models.Model):
 class List(models.Model):
     name = models.CharField(max_length=255)
     board = models.ForeignKey(Board, on_delete=models.CASCADE, related_name="lists")
-    position = models.PositiveIntegerField()
+    position = models.DecimalField(decimal_places=10, max_digits=20)
 
     class Meta:
         ordering = ["position"]
